DbScannerFixQMod3 - Updating to QModManager 3.0 and latest Harmony DLL - Moving everything into a single file
